add hook use statistics
